@@ -6,7 +6,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.6.3</version>
+        <version>2.7.4</version>
     </parent>
 
     <groupId>fr.pinguet62</groupId>
@@ -15,13 +15,13 @@
     <packaging>pom</packaging>
 
     <properties>
-        <java.version>17</java.version>
+        <java.version>18</java.version>
         <kafka.version>3.1.0</kafka.version><!-- fix Windows -->
     </properties>
 
     <modules>
         <module>asciidoc-site</module>
-        <module>embedded-elasticsearch-spring-boot</module>
+        <!--<module>embedded-elasticsearch-spring-boot</module>-->
         <module>jackson-undefined</module>
         <module>jsr303-validation</module>
         <module>other</module>
